+++ test/tempo/weekday_test.dart
@@ -1,24 +1,18 @@
 import 'package:tempo/tempo.dart';
 import 'package:test/test.dart';
 
-void main() {
-  test('us weekday numbers', () {
-    expect(Weekday.sunday.us, 0);
-    expect(Weekday.monday.us, 1);
-    expect(Weekday.tuesday.us, 2);
-    expect(Weekday.wednesday.us, 3);
-    expect(Weekday.thursday.us, 4);
-    expect(Weekday.friday.us, 5);
-    expect(Weekday.saturday.us, 6);
-  });
+Matcher isWeekday(int us, int iso) => isA<Weekday>()
+    .having((w) => w.us, 'us', us)
+    .having((w) => w.iso, 'iso', iso);
 
-  test('iso weekday numbers', () {
-    expect(Weekday.monday.iso, 1);
-    expect(Weekday.tuesday.iso, 2);
-    expect(Weekday.wednesday.iso, 3);
-    expect(Weekday.thursday.iso, 4);
-    expect(Weekday.friday.iso, 5);
-    expect(Weekday.saturday.iso, 6);
-    expect(Weekday.sunday.iso, 7);
+void main() {
+  test('weekday numbers', () {
+    expect(Weekday.sunday, isWeekday(0, 7));
+    expect(Weekday.monday, isWeekday(1, 1));
+    expect(Weekday.tuesday, isWeekday(2, 2));
+    expect(Weekday.wednesday, isWeekday(3, 3));
+    expect(Weekday.thursday, isWeekday(4, 4));
+    expect(Weekday.friday, isWeekday(5, 5));
+    expect(Weekday.saturday, isWeekday(6, 6));
   });
 }